feat: Add optional haptic feedback via react-native-haptic-feedback (#69)
## Description

Adds support for haptic vibrations when drag starts/ends or items order
changes. The library is used only when available (the
`react-native-haptic-feedback` is the optional dependency and haptics
are used only when the library exists).

The next PR will add support for `expo-haptics`, which is another
commonly used library for haptic feedback.