chore: Delay some operations after item touch (#139)
## Description

This PR just slightly reduces overhead of operations done when the touch
gesture is recognized but the press animation didn't start yet. There is
no point in performing these operations, such as changing layer zIndices
on item touch and these operations were moved to the timeout fired when
the press animation starts.