feat: Use layout animations for sortable items position updates (#172)
## Description

This PR replaces `withTiming` usage for all items position updates.
Instead, linear layout transition is used to automatically animate items
to their final positions.

The logic is more complex now but it is necessary as the previous
implementation didn't work on paper, where UI props are updated
synchronously and layout props asynchronously. After changes, everything
works as expected and should be even faster as it omits committing
changes to the shadow tree on Fabric.