chore: Add warning for debug and scrollableRef prop changes, adjust debug examples (#187)
## Description

`debug` and `scrollableRef` shouldn't be changed on the fly as they
add/remove providers in the component tree, which causes the entire
sortable component render. Because of that, the current state of the
component is lost.

To make it obvious for the user that they shouldn't modify these props
on the fly, after the component is rendered for the first time, I added
a warning for these props changes.