feat: Make drag activation timings and drop animation duration customizable (#226)
## Description

This PR exposes props changing the delay and duration of the item drag
activation. It also adds a separate prop for dragged item drop animation
duration customization (when the active item is released and animated to
its target position).

This PR also fixes bug that something occurred when an item was added to
the sortable component after initial render and the user started
dragging this item before any other layout change. In this case, the
item drop animation was immediate, which was an invalid behavior.