fix: View drop animation on native after adding shouldAnimateLayout (#237)
## Description

Small fix for the view drop animation on native after the
`shouldAnimateLayout` prop was introduced to fix page resizing behavior
on web. Before this change, the view appeared immediately on the target
position if no reordering happened and the view was dragged without
changing position from it initial render position.