fix: Remaining web implementation issues (#247)
## Description

Fixes remaining issues of the web implementation of sortable components.
Some things cannot be fixed on the library side and require setting
`overflowY` to `scroll` on the scrollable component or hiding the
scrollbar at all.

This PR also simplifies style calculation on the web as we aren't using
layout animations at all and everything is implemented with
transformations.