fix: Don't skip entering and exiting animations on user's views, improve layout animations docs (#271)
## Description

I use `LayoutAnimationConfig` to skip `itemEntering` and `itemExiting`
animation from being applied when the entire sortable component renders
or is unmounted. This is intentional but confusing, so I improved the
description in docs to make it clear how these props are supposed to be
used and how they work.

I also didn't unset `skipEntering` and `skipExiting` for views that the
user renders within the sortable component, which is an invalid behavior
preventing users from using any kind of `entering`/`exiting` reanimated
layout animation. To fix this issue, I simply added another
`LayoutAnimationConfig` that unsets `skipEntering` and `skipExiting`
(sets them to `false`).