feat: Smart drag callbacks handling on UI or JS thread (#326)
## Description

This PR adds smart drag callbacks handling. Before this change all
callbacks were executed on the JS thread, which is bad when they are
called very often (e.g. `onDragMove` is called on every little movement
of the active item).

Because of that I decided to implement smarter approach that will allow
users to decide on which thread the callback should be executed. If the
callback is marked as a `'worklet'`, then it is called from the UI
thread. Otherwise, callbacks are called from the JS thread.

It is very recommended to mark callbacks as worklets when possible to
avoid calling `runOnJS` and jumping between threads. It should result in
a better performance.