chore: Improve callbacks example (#327)
## Description

This PR improves callbacks example screen by adding a possibility to
select between a worklet function callback executed on the UI thread and
the JS callback function run from the JS thread. I also added a
possibility to disable specific callback functions (as the recently
added `onDragMove` can makes the `onOrderChange` callback value hard to
see as it is executed on every little movement of the active item).

## Changes showcase


https://github.com/user-attachments/assets/3b86dbf8-7767-4db8-9f95-bd06d45d4e4c