chore: Add react-require plugin, patch bob to preserve JSX (#328)
## Description

This PR introduces some changes to the build script. It patches the
`react-native-builder-bob` library to preserve JSX syntax in generated
JS files. It also adds a `babel-plugin-react-require` babel plugin to
add `React` imports to all files where JSX is used (the import shouldn't
be needed but is added just in case to prevent `React is not defined`
error from being shown in projects with not valid configuration).