feat: Delayed absolute layout initialization before sorting is enabled for the first time (#329)
## Description

This PR adds delayed absolute layout initialization when the
`sortEnabled` prop os the sortable component is set to false. It reduces
the number of calculations and applied animated styles to the minimum to
make the initial render of the sortable component less demanding and
calculates and applied the absolute layout only after the `sortEnabled`
was set to `true` for the first time.