fix: Remove react-native-builder-bob patch (#339)
## Description

This PR removes a `react-native-builder-bob` patch which was intended to
generate js files with unprocessed jsx syntax. This may be problematic
for test runners, like jest, which may expect a correct JS syntax
(without jsx).

As a result, I decided to revert it back to the previous implementation
without patch.