chore: Move active item related value to a separate provider (#359)
## Description

This PR is just a preparation for the next one in which I will add
support for item dragging between sortable components (like dragging
from one grid to another). To implement this feature, I need to know the
current position of the active item in all sortable components to detect
when it enters another sortable component.