fix: Invalid custom handle measurement (#384)
## Description

This PR fixes issues caused by recent changes introduced in the #377 PR.

I modified the gesture in the custom handle component by adding the
`onBegin` method to the gesture. This didn't work when the custom handle
was rendered after the item (e.g. with some timeout). This PR fixes this
invalid behavior and simplifies handle measurement logic a bit.