fix: Multi touch flickering issue (#392)
## Description

This PR fixes multi touch flickering issue caused by the `handleDragEnd`
being called on the cancelled gesture. The second touch gesture was
properly ignored and failed but it called `handleDragEnd` as this method
is always called in the `onFinalize` callback, which invalidated not
only the second gesture that we wanted to ignore, but also the first one
that was correct and should be still handled.

The behavior on Android is unfortunately not perfect, because when we
start touching a second item, the gesture on the first one gets
invalidated as the `onTouchesUp` callback is not properly called on the
gesture instance. This is a `react-native-gesture-handler` issue that I
reported here:

https://github.com/software-mansion/react-native-gesture-handler/issues/3543