perf: Replace useSharedValue with useMutableValue (#408)
## Description

This PR replaces all `useSharedValue` usages with `useMutableValue`.
`useSharedValue` is heavier because of the cleanup `useEffect` that
cancels currently running animation if there is any:

```
export function useSharedValue<Value>(initialValue: Value): SharedValue<Value> {
  const [mutable] = useState(() => makeMutable(initialValue));
  useEffect(() => {
    return () => {
      cancelAnimation(mutable);
    };
  }, [mutable]);
  return mutable;
}
```

I don't need this behavior as I never use infinite animations and, in
most cases, don't use animations at all, rather update value by myself.
I only use animations for items reordering and the drop indicator, but
since they are just short `withTiming` animations, the cleanup is not
necessary. Even if the animation keeps running after the view was
unmounted, it runs for a brief while, so there is no performance benefit
in eager cleanup and it's better to reduce the number of unnecessary
`useEffect` hooks.