feat: New auto scroll implementation (#454)
## Description

This PR is a complete rewrite of the auto-scroll feature that considers
limitation of the New Architecture (scroll performance issues reported
in the #285), as well as, requirements for the new collapsible items
feature (#453).

The new auto scroll implementation uses better, time-based scroll
distance calculation, includes a possibility to reduce the `scrollTo`
calls frequency to prevent app from freezing on Android (see #285) and
has more customization props.