fix: Active item touch start position on Android (#462)
## Description

This issue was caused by incorrect `measure` function output on Android
(the same incorrect output is returned by the Reanimated's and RN's
implementation). This was best visible after scrolling a bit in the
collapsible items example, pressing the item once and pressing it again
without changing the scroll position. At the second press, the
container's `pageY` position was completely wrong, causing the active
item to fly away far from the user's finger.

As a workaround, I decided to calculate the item's position in the
container's position independent way.