feat: Data provider (#472)
## Description

This PR adds more performant way of sortable component items rendering.
It triggers render only for items that changed instead of going through
the entire providers tree from the top to the bottom to re-render an
item.

This also will make it possible to change the data in the multizone
provider when the item enters another zone (the state can be now
controlled from another provider, not only changed via props).