feat: Better collapsible auto-scroll behavior (#457)
## Description

This PR is a huge improvement over the initial collapsible items
implementation introduced in the #444 PR. It re-implements the core
logic so that updates are performed synchronously with layout changes,
thanks to which there are far less glitches and unexpected actions.

It also implements a smart way of scroll offset restoration after items
expand to their normal positions.

## Changes showcase


https://github.com/user-attachments/assets/2cb3ab86-d6a5-412b-b46e-6f2e80a2152e

## TODO

- [x] Fix horizontal collapsible items layout positions animation after
collapse
- [x] Test if it works on Paper