fix: Autoscroll exceeding bounds (#473)
## Description

This PR fixes the issue caused by the fact that `measure` sometimes
returns slightly outdated values, thus the container scroll offset was
invalid and we were scrolling beyond the bounds of the scrollable
content size.

To fix the issue, I decided to check whether the new `contentOffset` is
different from the old one before calling the `scrollBy` function in
order not to queue too many scroll updates, which can sum up to the
invalid distance outside of the scrollable content container size.